requirements.txt
fastapi
uvicorn
langchain
openai
pdfplumber
python-docx
beautifulsoup4
aiohttp
python-dotenv
pymysql
sqlalchemy